.testing/requirements-dev.txt
# Development dependencies for cc-plugins testing
# Install with: pip install -r requirements-dev.txt

# Testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-timeout>=2.0.0

# Mocking and test utilities
responses>=0.23.0
pytest-mock>=3.0.0

# Environment management
python-dotenv>=1.0.0

# Type checking (optional)
mypy>=1.0.0

# Code formatting (optional)
black>=23.0.0
isort>=5.0.0

# Linting (optional)
flake8>=6.0.0
